Further breakout of how commands are added

+++ index.js
@@ -7,19 +7,22 @@ dotenv.config();
 //This loads all data models at once
 const models = require('./models');
 
-//Load our handlers
+//Load our edit handler
 const editHandler = require('./handlers/edit')(models);
-const commandHandler = require('./handlers/commands')(models);
+
+//Pull in all command handlers at once
+const handlers = require('./handlers/commands')(models).handlerList;
 
 client.login(process.env.BOT_TOKEN);
 
 client.once('ready', async () => {
     await models.sequelize.sync();
     console.log('Ready!');
+    await client.channels.find(x => x.name === process.env.CHANNEL_NAME).send("Who's afraid of being banned?");
 });
 
 client.on('messageUpdate', async (oldMsg, newMsg) => {
-    if (newMsg.channel.name == process.env.CHANNEL_NAME && newMsg.author.id != client.user.id)
+    if (newMsg.channel.name == process.env.CHANNEL_NAME && newMsg.author.id != client.user.id && oldMsg.cleanContent != newMsg.cleanContent)    //cleanContent check should prevent deletes caused by discord embedding an image
     {
         await editHandler.deleteAndInsult(newMsg);
     }
@@ -27,16 +30,14 @@ client.on('messageUpdate', async (oldMsg, newMsg) => {
 
 client.on('message', async msg => {
     if(msg.channel.name == process.env.CHANNEL_NAME && msg.author.id != client.user.id) {
-        var user = await models.Stamps.findOne({where: { username: msg.author.id }});
+        var user = await models.Stamps.findOne({where: { discordId: msg.author.id }});
         var firstTime = false;
         if(!user) {
-            user = await models.Stamps.create({username: msg.author.id, timestamp: msg.createdTimestamp + parseInt(process.env.DEFAULT_COOLDOWN)})
+            user = await models.Stamps.create({discordId: msg.author.id, tag: msg.author.tag, timestamp: msg.createdTimestamp + parseInt(process.env.DEFAULT_COOLDOWN*1000)})
             firstTime = true;
-            console.log("New user created. Msg timestamp and wait timestamp are");
-            console.log(msg.createdTimestamp);
-            console.log(user.timestamp);
         }
         if (!firstTime && msg.createdTimestamp < user.timestamp) {
+            await msg.author.send("Your message was deleted. You need to wait " + ((user.timestamp - msg.createdTimestamp)/1000).toString() + " more seconds.\n > " + msg.content);
             await msg.delete();
         } else {
             user.timestamp = msg.createdTimestamp + user.cooldown;
@@ -60,14 +61,34 @@ client.on('message', async msg => {
                 const command = input.shift();
                 const commandArgs = input.join(' ');
                 
-                //This is the block to add your new commands
-                switch(command){
-                    case 'setCooldown':
-                        await commandHandler.setCooldown(user, msg, commandArgs);
-                        break;
-                    case 'addInsult' :
-                        await commandHandler.addInsult(msg, commandArgs);
-                        break;
+                if (command.toLowerCase() == 'help'){
+                    if (commandArgs == "") {
+                        helpMessage = "Here are all of the ranting commands I accept. You call them via `![commandName] [arguments]`. To get more info, type `!help [commandName]`\n";
+                        handlers.forEach( handler => {
+                            helpMessage += handler.description() + "\n";
+                        });
+                        await msg.channel.send(helpMessage);
+                    } else {
+                        helpMessage = ""
+                        handlers.forEach(handler => {
+                            if (handler.name.toLowerCase() == commandArgs.toLowerCase())
+                                helpMessage = handler.help();
+                        });
+                        if (helpMessage != "")
+                            await msg.channel.send(helpMessage);
+                        else
+                            await msg.channel.send("\"" + commandArgs + "\" does not match any of the commands I know");
+                    }
+                } else {
+                    handlerFound = false;
+                    await Promise.all(handlers.map( async (handler) => {
+                        if (handler.name.toLowerCase() == command.toLowerCase()){
+                            await handler.execute(user, msg, commandArgs);
+                            handlerFound = true;
+                        }
+                    }));
+                    if (!handlerFound)
+                        await msg.channel.send("\"" + command + "\" does not match any of the commands I know");
                 }
             }
         }
